// src/domain/models/Student.model.ts
export interface IStudentsModel {
  id?: number;
  idDocumentsType: number;
  idHeadquarter: number;
  names: string;
  lastName: string;
  age: string;
  birthday: string;
  address: string;
  tutorNames: string;
  tutorLastName: string;
  cellphoneTutor: string;
  documentNumber: string;
  email: string;
  schooling: string;
  grate: string;
  school: string;
  diagnosis: boolean;
  medicalReport: boolean;
  certificateDisability: boolean;
  conadisCard: boolean;
  psychologicalReport: boolean;
  state: boolean;
}